Do not check for changes in unrelated charts
The chart_version_check.sh code checks if modified files are in a
directory whose chart version has changed. Until now, changes in any
of the voltha-* directories required a version bump in voltha/ as well.

This does not appear to be intentional. It results in output like this:

-------------------- %< -----------------------------------------
[CHART] voltha-tracing/Chart.yaml                                    (1.0.3 => 1.0.4)
    Files Changed: voltha-tracing
        voltha-tracing/values.yaml
[CHART] voltha/Chart.yaml
    Files Changed: voltha
        voltha-tracing/values.yaml
[ERROR] Chart modified but version unchanged: voltha
-------------------- %< -----------------------------------------

This commit changes the check to require an exact match of the directory
name.

Change-Id: Ib74d9f8eb63a054dc9c9bcff270433db892c8a3b
Signed-off-by: Roger Luethi <roger.luethi@bisdn.de>